backend/requirements.txt
# Web Framework
fastapi==0.109.0
uvicorn[standard]==0.27.0
python-multipart==0.0.6

# Database
sqlalchemy==2.0.25
psycopg2-binary==2.9.9
alembic==1.13.1

# Redis
redis==5.0.1
hiredis==2.3.2

# AI & ML
openai==1.10.0
tiktoken==0.5.2
zhipuai>=2.1.5  # 智谱 AI SDK

# Data Processing
pydantic==2.6.0
pydantic-settings==2.1.0
python-dotenv==1.0.0

# Authentication & Security
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
bcrypt==4.1.2

# Utils
httpx==0.26.0
python-dateutil==2.8.2

# Development
pytest==7.4.4
pytest-asyncio==0.23.3
black==24.1.1
flake8==7.0.0

